// packages/twenty-server/src/engine/core-modules/managed-email/enums/managed-email-infrastructure-state.enum.ts
export enum ManagedEmailInfrastructureState {
  AWAITING_PAYMENT = 'AWAITING_PAYMENT',
  ORDERING = 'ORDERING',
  PROVISIONING_DOMAIN = 'PROVISIONING_DOMAIN',
  PROVISIONING_MAILBOX = 'PROVISIONING_MAILBOX',
  WAITING_FOR_CREDENTIALS = 'WAITING_FOR_CREDENTIALS',
  CONNECTING_TWENTY = 'CONNECTING_TWENTY',
  ACTIVE = 'ACTIVE',
  PAYMENT_REQUIRED = 'PAYMENT_REQUIRED',
  DEACTIVATING = 'DEACTIVATING',
  INACTIVE = 'INACTIVE',
  REPLACEMENT_REQUIRED = 'REPLACEMENT_REQUIRED',
  RECONCILIATION_REQUIRED = 'RECONCILIATION_REQUIRED',
}