// packages/twenty-server/src/engine/core-modules/managed-email/enums/managed-email-warmup-state.enum.ts
export enum ManagedEmailWarmupState {
  NOT_APPLICABLE = 'NOT_APPLICABLE',
  CONNECTING = 'CONNECTING',
  WARMING = 'WARMING',
  MAINTENANCE = 'MAINTENANCE',
  CANCEL_AT_PERIOD_END = 'CANCEL_AT_PERIOD_END',
  PAUSED = 'PAUSED',
  ACTION_REQUIRED = 'ACTION_REQUIRED',
  DELETING = 'DELETING',
  DELETED = 'DELETED',
  RECONCILIATION_REQUIRED = 'RECONCILIATION_REQUIRED',
}